requirements-train.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
scikit-learn==1.5.2
xgboost==2.1.3
pandas==2.1.4
numpy==1.26.4
joblib==1.4.2
sqlalchemy
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.3
kaggle==1.6.17
apache-airflow==3.0.0